CSS Image-Hover Effect
A hover effect is an animation that triggers when a user hovers over an element on the page. A CSS hover effect takes place when a user hovers over an element, and the element responds with transition effects. It is used to mark the key items on the web page and it's an effective way to enhance the user experience.